Former-commit-id: e68f84ed9fb717fc8e53d28a170164b157f4a608

+++ openquake/commonlib/readinput.py
@@ -520,7 +520,8 @@ def get_source_models(oqparam, gsim_lt, source_model_lt, monitor,
             logging.info('%s has been considered %d times', fname, hits)
             if not psr.changed_sources:
                 dupl += hits
-    if dupl and not oqparam.optimize_same_id_sources:
+    if (dupl and not oqparam.optimize_same_id_sources and
+            'event_based' not in oqparam.calculation_mode):
         logging.warn('You are doing redundant calculations: please make sure '
                      'that different sources have different IDs and set '
                      'optimize_same_id_sources=true in your .ini file')
@@ -553,6 +554,9 @@ def get_composite_source_model(oqparam, monitor=performance.Monitor(),
     idx = 0
     gsim_lt = get_gsim_lt(oqparam)
     source_model_lt = get_source_model_lt(oqparam)
+    if oqparam.number_of_logic_tree_samples == 0:
+        logging.info('Potential number of logic tree paths = {:,d}'.format(
+            source_model_lt.num_paths * gsim_lt.get_num_paths()))
     if source_model_lt.on_each_source:
         logging.info('There is a logic tree on each source')
     for source_model in get_source_models(
@@ -970,7 +974,7 @@ def get_mesh_hcurves(oqparam):
 
 
 # used in utils/reduce_sm and utils/extract_source
-def reduce_source_model(smlt_file, source_ids):
+def reduce_source_model(smlt_file, source_ids, remove=True):
     """
     Extract sources from the composite source model
     """
@@ -986,8 +990,11 @@ def reduce_source_model(smlt_file, source_ids):
             for src_group in origmodel:
                 sg = copy.copy(src_group)
                 sg.nodes = []
-                weights = src_group['srcs_weights']
-                assert len(weights) == len(src_group.nodes)
+                weights = src_group.get('srcs_weights')
+                if weights:
+                    assert len(weights) == len(src_group.nodes)
+                else:
+                    weights = [1] * len(src_group.nodes)
                 src_group['srcs_weights'] = reduced_weigths = []
                 for src_node, weight in zip(src_group, weights):
                     if src_node['id'] in source_ids:
@@ -1000,7 +1007,7 @@ def reduce_source_model(smlt_file, source_ids):
             with open(path, 'wb') as f:
                 nrml.write([model], f, xmlns=root['xmlns'])
                 logging.warn('Reduced %s' % path)
-        else:
+        elif remove:  # remove the files completely reduced
             os.remove(path)
